evo-prop-sso/requirements.txt
anyio==4.4.0
asgiref==3.8.1
async-property==0.2.2
boto3==1.34.159
botocore==1.34.159
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
deprecation==2.1.0
django==4.2.5
django-cors-headers==4.4.0
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jmespath==1.0.1
jsonfield==3.1.0
jwcrypto==1.5.6
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-keycloak==4.3.0
requests==2.32.3
requests-toolbelt==1.0.0
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
typing_extensions==4.12.2
uvicorn